--- src/Data/Utils/Utils.java
@@ -1,20 +1,0 @@
-package Data.Utils;
-
-import java.text.DecimalFormat;
-import java.text.DecimalFormatSymbols;
-
-public class Utils {
-    public static double convertDoubleToTwoDecimals(double x, int decimalPlaces) {
-        DecimalFormatSymbols symbols = new DecimalFormatSymbols();
-        symbols.setDecimalSeparator('.');
-
-        StringBuilder pattern = new StringBuilder("#.");
-        for (int i = 0; i < decimalPlaces; i++) {
-            pattern.append("#");
-        }
-
-        DecimalFormat dfConverted = new DecimalFormat(pattern.toString(), symbols);
-
-        return Double.parseDouble(dfConverted.format(x));
-    }
-}
