+++ src/Utils/Data/Calculations.java
@@ -1,6 +1,6 @@
-package Data;
+package Utils.Data;
 
-import Data.Utils.Utils;
+import Utils.Utils;
 
 public class Calculations {
     public static void convertCurrencies(String baseCur, String targetCur, double amount) {
@@ -9,7 +9,7 @@ public class Calculations {
         if (ExchangeRateFetcher.latestRate != 0.0) {
             System.out.println("\n========================================================================\n" + 
             amount + " " + baseCur + " sind " +
-            Utils.convertDoubleToTwoDecimals(amount * ExchangeRateFetcher.latestRate, 2) + " " + targetCur + 
+            Utils.adjustDecimal(amount * ExchangeRateFetcher.latestRate, 2) + " " + targetCur + 
             " bei einem Wechselkurs von " + ExchangeRateFetcher.latestRate + 
             "\n========================================================================");
